Update flow for single and self-employed mothers
For single parents without the minimum weeks worked there is no
need to also present them with the minimum earnings question as both
of these conditions must be positive to qualify for allowance.

Instead the single_birth_nothing outcome can be presented.

If a self-employed mother has answered "yes" to working 26 weeks there is
no need to also ask if they have earned £390. This change skips this question inthis case, directing to mat_allowance if single or to the partner questions.

It also adjusts a method on the calculator to not examine the unanswered question
for self-employed mothers, as this can be used in later calculations if the
mother has a partner.