Extend HMRC C-19 late payment penalties to 1st Apr
What

Currently, the trigger date for self assessment late payment penalties
for 2019-20 is 2 March. Because of coronavirus, HMRC have decided to
extend it to 1 April. This means we need to change the calculation
formula, so that users inputting a date between 2 March and 1 April
don't have a late payment penalty listed on their outcome.

Why

If we don't do this there's a risk that some users will think they need
to pay a late penalty payment when they don't. The late payment penalties
are quite a lot of money, so this could cause unnecessary stress for
users at a time when many may already be having financial problems. There's
also a reputational risk to HMRC and GOV.UK, and a risk to our relationship
with HMRC.