Remove GitHub Actions from CI checks
This was started with good intentions but the work has stalled.
Running CI on both GitHub Actions and Jenkins is pointless and
slows down PRs - and has caused issues with rolling out urgent
content updates. (The developer ended up disabling GH actions
temporarily).

This work is now likely to be considered in the Replatforming
team.

Trello: https://trello.com/c/IvaSyaIz/150-switch-smart-answers-to-use-github-actions-for-ci