Rename SessionAnswersController to FlowController
The naming "SessionAnswers" is somewhat confusing as it's used uniquely
here and isn't defined anywhere. It orginally was named this to
represent a SmartAnswer that used session for handling user responses.
However this controller will in future not only support sessions, but
also query string. This generalises the name and uses terminology we
have defined.