Move 'Start again' link
- Separate 'Start again' link from 'Your answers' section as part of design interation.

The amount of assertions which test the existence of the “Start again” link has been reduced.
Making this assertion once per test file should be enough given that there is nothing that
differs between the tests that would effect how the link displays for each test case.
The functionality of the Start Again link is also tested in:
 spec/features/smart_answers/flow_navigation_spec.rb.

The `caption` element has been removed as it is an accessibility concern that the header
and link weren't found by screenreaders:
https://trello.com/c/dtbrSntn/586-accessibility-smart-answers-fix-link-in-table-caption