Allow redirect type and segments_mode in unpublish redirects
By default Publishing API redirects will have a segments_mode of
"preserve". This means that, for a prefix route of /item-a, to /item-b
that a request of /item-a/a/b/c will redirect to /item-b/a/b/c. This is
rarely what we want as Smart Answers typically have a lot of potential
segments. To resolve this segments_mode is set to "ignore".

I also noticed that the two routes registered for Smart Answers are
different types, there is an exact route for a start page and a prefix
route for a flow page. To allow these to be accurately redirected I've
added a type parameter and updated the examples to use "exact" for start
pages.