Back button functionality for url and session based flows
For session based flows, the back button navigation is determined by the
previous_state.node_slug.

For url based journeys we cannot determine the previous state as cleanly as we can
with the session based approach. The destination of the back link is therefore
determined based on the responses and question number.