Set the unpublish tasks to discard drafts
By default the Publishing API will refuse to update the unpublishing of
content which has a draft. This can be problematic if we've temporarily
taken a Smart Answer offline and marked it as a draft.

This can be resolved by telling the Publishing API that we don't mind
them discarding the draft. This feature is normally in place so editors
draft work isn't lost, however since the draft is all part of this repo
we don't need to precious about it.