Fix text on question
Deductions in the child benefit tax calculator need to be spilt based on 
whether they are grossed up or not. 

We accidently stated that gift aid should not be included in a question 
where it should be included. This fixes the text to state that 
retirement annuities should be excluded rather than gift aid.