Add `/flow/` prefix routes
We would like to migrate the existing `/s/` routes to use the new
`/flow/` prefix. This is to reduce confusion around this extra path slug
that is required due to technical limitation with GOV.UK's router.
"flow" is more descriptive than "s". "s" was originally thought to
mean "session" when session based flows were introduced, however we'd
like to generalize the FlowController to also support storing user
responses in the query string.