Fix issue using fixture flows in rspec tests
This addes before and after hooks to reset the flow registry between the
fixture directory and the existing flow directory. This allows us to use
the fixture directory to write dummy flows for testing the core
functionality of Smart Answers with out depending of existing flows that
may change and break tests.