Add redirect for path based flows if accessed with flow prefix
The new `/flow/` prefix routes will be registered for all flows,
including the existing path based flows. This means a request for a path
based flow may be handed off to the FlowController. This adds a blanket
check to the controller to redirect request for path based flows back to
the original routes. We should rarely get this type of request as the
existing flow start pages and flow navigation use the old url paths.
However, this just catches an edge cases if a user did manually use the
`/flow/` in the URL.