Code in terms of outcomes, rather than answers
Following the logic of nested conditionals requires keeping track
of lots of information. It can be very difficult to give an answer
to the question "how do I reach this outcome?". It can also be
difficult to deduce whether multiple conditions can lead to the
same outcome, and what those conditions might be.

Instead of calculating in terms of the user's answers, it is much
simpler to treat each outcome in turn, and calculate whether the
user's answers should qualify for that outcome. Advantages:

- Flattens the logic structure, so that we have one long if-else
  block, rather than a heavily nested (complex) structure.
- The conditions for each outcome are defined in one place, rather
  than multiple paths of code being able to lead to the same
  outcome.
- It's now much easier to explain how each outcome is arrived at.