Move the COMMONWEALTH_COUNTRIES list so it can be shared
The marriage-abroad list is actually not a complete list, so I've created a complete list then removed the ones to which marriage-abroad does not apply: Australia, Gambia, Kenya, Maldives, Mozambique, Rwanda, South Africa, and Tanzania. Feasibly there are reasons for all of these to be special cases, but it seems to make sense to have the commonwealth-countries list actually reflect the real membership.