Set up logic for new schemes in outcomes
This moves VAT reduction to be above the council section and applies the
logic to it - to reflect that this isn't a council grant.

It changes the conditional for council grants to use the new shorter
one - In the next commit I will adjust that further but I wanted to
reduce whitespace changes here.

It moves LRSG Closed Addendum up to match the point in the spreadsheet
and corrects the logic to used the lrsg_closed_addendum.

Finally it moves the restart grant into the council section at the
bottom.