Fetch company name from Companies House API
For the "Next steps for your business" flow we use the company name in
the title of the results page. This  derived from the supplied company
registration number to query the Companies House Public Data API.
Displaying the company name isn't critical, so there is a fallback
heading so user's continue to see results if the request fails.