Add start page content to hidden search terms
The body and post body content of start pages are used by the Search API
to index transaction pages. However, that content is no longer
represented in the new smart_answer schema as the
`introductory_paragraph` and the `more_information` fields were removed.
This adds the content to the hidden search terms by expanding the
extract_flow_content method to include the start page body and post body
content.