Remove reference to `outcome_no_visa_needed` outcome
This outcome is no longer used or existing but is still referenced in
the flow definition. This doesn't cause issue for the flow as not used
in the user journey. However prevents the flow from being published to
the Publishing API, as the content item presenter will attempt to
extract content for that outcome node and fail as the template file is
missing.