Remove /flow/ prefix routes
These are no longer needed as Smart Answer's is now represented by a
single content item. Therefore the extra slug is not required to
get around the constraint where the two content items had to have
seperate base paths. Instead of renaming the the /s/ to /flow/ which was
original intended, the can remove the slug altogether.