Add `content_item` method to FlowHelper
This changes access to the content item from an instance variable to a
method call. This allows the removal of code with in the controller to
pre-load the content item, which is fragile is it needs to be aware
which templates it needs used. Using a memoised method will allow access
to content item where ever needed in the views, however only request the
content item if its actually used.

This also fixes a bug in which content items weren't previously being
retrieved for questions in session based flows. This meant that question
views didn't render the metatags component necessary for prevent dates
and postcodes from being sent to GA. This example of the fragility
mentioned above.