Re-instate the self_employed_income_scheme question
This question is being returned to the Business Support Finder as the
Self-employed Income Scheme is returning.

Previously we had a check about whether a business had 0-249 employees
before deciding whether to show this question, however the documentation
no longer has this so I've not included it. I think it could be that it
makes no sense to be self employed and have 249+ employees, but in that
case we probably just shouldn't show the question?

I've included the copy as it was on the self-employed question with the
clause about holding a limited company directorship, this should be checked.