Get current response for session based flows
There is currently a bug in session-based smart-answers, where answers
to previous questions are not being populated even though the answers
exist in the session.

This change gets the existing answer for session-based flows.

The information about the current state and existing responses can be
easily accessed in FlowPresenter, and as all of the
answers are populated by one of many QuestionPresenters and they
all have access to the FlowPresenter, it seemed best to gather the
responses here.

Getting the responses in one place (`FlowPresenter`) means that in
future commits we can remove the `prefill_value_*` methods in the helper
and instead have each `QuestionPresenter` get the value of the current
question from `FlowPresenter` and have the presenter itself decide how
it needs to work with the data in custom cases, e.g. checkboxes.

The benefit of this approach is that the presenter doesn't need to know
what type of flow (url, session) the smart-answer is.

We cannot use the existing `accepted_responses` or
`normalize_responses_param` methods as these must return an array so
that they can be used to populate the "Your answers" section at the
bottom of each page.