Update results for Next steps for your business flow
This updates the results files for include the id and groups attributes.
The ID is needed to reference filtering rules that are yet to be added.
The group specificies the top level group title for the result. This
also renames "section_name" to topic, as it's more concise. This also
remove to results that are no longered need and updates the descriptions
for two others.