Replace `prefill_value_includes?` with `checked?`
To replace the `prefill_value_includes?` we need to update the tests to
stub the response from `response_for_current_question`. This is because
`checked?` depends on `response_for_current_question` being available to
call.

`response_for_current_question` is delegated from `FlowPresenter`,
however, for ease, in the tests, a `nil` FlowPresenter object is being
passed to the presenter, so the presenter does not have access to any
its methods. Rather than stubbing an entire FlowPresenter object, we can
instead stub the response from the only method we need for the tests to
pass.