Replace `prefill_value_is?` with `selected?`
Currently pre-population of radio button questions using the `prefill_*`
methods in the helper only work for url-based flows, not session-based
flows whereas `response_for_current_question` returns the current value
for both.

This uses the `response_for_current_question` method added in PR #5315
to get the existing response for checkbox questions.

To replace the `prefill_value_is?` we need to update the tests to
stub the response from `response_for_current_question`. This is because
`selected?` method depends on `response_for_current_question` being
available to call.

`response_for_current_question` is delegated from `FlowPresenter`,
however, for ease, in the tests, a `nil` FlowPresenter object is being
passed to the presenter, so the presenter does not have access to any
its methods. Rather than stubbing an entire FlowPresenter object, we can
instead stub the response from the only method we need for the tests to
pass.