Get date values from response_for_current_question
Currently pre-population of date questions using the `prefill_*` methods
in the helper only work for url-based flows, not session-based flows
whereas `response_for_current_question` returns the current value for
both.

This uses the `response_for_current_question` method added in PR #5315
to get the existing response for date questions.

To replace the `prefill_value_*?` we need to update the tests to
stub the response from `response_for_current_question`. This is because
`selected_*` methods depend on `response_for_current_question` being
available to call.

`response_for_current_question` is delegated from `FlowPresenter`,
however, for ease, in the tests, a `nil` FlowPresenter object is being
passed to the presenter, so the presenter does not have access to any
its methods. Rather than stubbing an entire FlowPresenter object, we can
instead stub the response from the only method we need for the tests to
pass.

`response_for_current_question` calls the `to_response` methods for Date
if `previous_response` is in the query string. This `to_response`
method returns `nil` if `previous_response` is not a valid date, so we
don't need to add any extra guards for this scenario in the `selected_*`
methods.