Do question formatting in FlowPresenter
In the CurrentQuestionHelper methods, that
`response_for_current_question` is replacing, the output from
`params[:previous_response]`. This is because
`params[:previous_response]` is a querystring parameter and there for
everything in it is a string.

In the case of checkboxes, that means the values appear as a comma
separated string rather than as an array. To account for this there is
already a `to_response` method on each question node that returns the
appropriate object type. So for checkboxes, `to_response` takes a comma
separated string and returns an array of strings, but for other types of
question, e.g. money, in input is unchanged. This means that it is safe
to apply the  formatting so that it does not need to be replicated in
the other  QuestionPresenters in later PRs when the other
CurrentQuestionHelper methods are replaced.