Remove nightclub exemption from business support finder
The closed_by_restrictions outcomes are no longer exclusive to
businesses that are not nightclubs. Nightclubs were previously not shown
the LRSG Closed Addendum which we believe they could be eligible for.

We discussed whether to remove the nightclubs_or_adult_entertainment
option as no results depend on it, however content wanted to keep this
as it provided a distinction with leisure which nightclubs are typically
not eligible for.