Move node_slug from NodePresenter to Node
This method seems unconventional on a NodePresenter as other name
manipulation concerns are on the Node itself. I also altered this to
drop any trailing question mark that is common on Smart Answers by using
the exisiting filesystem_friendly_name method

Doing this work allows unblocking problems I've been having where a node
is readily available but the slug is not without the presenter which
means coupling the look-up of a path with the calculations of the
current Smart Answer state.