Decouple FlowController#start from flow state
At the start of a flow a user should always be going to the first
question, so rather than use aspects of the current state of a Smart
Answer we can use the slug of the first question.

I've done this because I find the next_node_slug a bit misleading
(whether it's actually next is contextual - it's better understood as
current) and it has it's own custom slug system.