Guard against invalid salary in params[:response]
An error can occur if `params[:response]` contains data in an invalid format.

The "period" field is a drop-down list with pre-defined values, so it doesn't benefit
the user to return the existing value. However "amount" is a text field, so returning
the value here allows them to see whatever they previously entered.