Remove tests for ensuring flows are visualiable
These tests ensured that every flow was visualisable. However, somewhat
pointless because if a flow is not visualisable, you cannot neccesarily
fix it due to the limitation of the graphing logic. In which case, the
flow would be added to the exemption list. This is somewhat limited in
value as there isn't a way to overcome a failed test.