Remove tests for content id for every flow
These tests ensure there is a content id for each flow. It is
inconsistent to have checks for this attribute and not others e.g.
title, name etc. This is essentially putting a constraint of the data,
rather than testing the functionality of the code. It would be better to
modify the `Flow` class to check that a content_id is present and test
that functionality.