Store start_node on Flow
This moves the creation of start_node to be a Flow concern rather than a
FlowPresenter concern. The reason for this is to allow usages of a
start_node without artificially creating a presenter with fake state.
Logically it makes sense that the start_node is something a flow knows
about.

This doesn't map perfectly to a flow though which has a concept of a
@node collection that this doesn't map to directly (a start node is
always needed but can't be created until a name is known). Because of
this the method creates a new node each time rather than doing any
memoing.