Do not suppress "none" option in `to_response`
The `to_response` method is used to format existing responses to a
question so that they can be pre-populated on a question page.

The `to_response` field was removing the "none" option, which meant that
in url-based flows, if a user selected "none" in a checkbox question,
and then went back to change their answer, nothing would be populated,
and it would appear as though the user had never answered at all.
This issue doesn't seem to affect session-based smart-answers.