Make to_response for checkboxes more resilient
Each question (aka "node") has its own `to_response` method. That means
we don't know which `to_response` method is going to be called until
runtime. That means we need to make sure each of these methods is
resilient enough to handle being passed invalid data without causing the
application to fail.

This guard is more of a precaution as it is very unlike that
`to_response` will ever have to handle anything other than strings or
arrays neither of which cause the `split` method to fail.