StateResolver class to populate State objects
This class has been added to abstract the concerns of constructing a
state object to a separate class. This is to allow for the removal of the
state management code in Flow and FlowPresenter.

The state created by this class is more elaborate than the ones of the
past as it aims to better capture the needs of understanding what a user
has answered for a current question and has the data to support
navigating with past questions. This has been done to remove the need
for extraneous code to state for dealing with any past responses [1]

One thing that seems to be a bit sub-optimal here is the use of a next
transition in apply_response_to_state, this feels a bit clunky but seems
the only abstraction available to determine if an answer is an error. To
resolve this would likely involve digging further into changing the
transition/state code which would probably be quite slow to do.

[1]: https://github.com/alphagov/smart-answers/blob/2bba0fba144240641577a88e6e88bd537eb133b6/app/presenters/flow_presenter.rb#L81-L91
[2]: https://github.com/alphagov/smart-answers/blob/2bba0fba144240641577a88e6e88bd537eb133b6/lib/smart_answer/flow.rb#L142