Remove `/s/` prefix for default routes to FlowController
This sets the defaults routes to the FlowController to not be prefixed
with `/s/`. The prefix was introduced to deal with a constraint of
publishing the start page and the flow as seperate content items in the
Publishing API. The two content items needed unqiue base paths. However,
Smart Answers now render the start page and are represented by a single
content item and this constraint is removed. Removing the prefix
simplifies the url structure and removes confusion around the meaning of
the prefix.

The old prefix routes will continue to be supported until
existing traffic no longer uses them - then they can be removed
entirely. This we automatically update the urls in the pages, so users
who continue their user journey will begin to use the non-prefixed
routes.