Refactor question response GA tracking
For each question we sent GA event to track user responses for certain
question types. This was original set up to use the question id which
was present in the `data-question-key` attribute on the form element.
However the attributes value was being set to the question title rather
than question id. Therefore previous tracking has references the
question title, not the question id.

This fixes the tracking method to use the actual h1 title of the page to
be more accurate and allow updating the values of `data-question-key` to
be the question key without breaking existing analytics.