Add response sumbission GA tracking with question id
Previously response submission GA events used the question text, instead
of the question key. This makes the tracking fragile when the question
text changes, but is still fundementally the same question. Now an extra
GA events will be sent with the question key instead of the question
text. This will allow analysis of tracking data using the quesion keys,
whilst retaining old events to support previous tracking analysis. In
future, it would be ideal to remove the old extra GA event with the
question text.