Fix error messages when no answer for a question
There was a bug with query parameter based flows where no error message
was displayed when a user tried to progress without selecting a option
on a question. This was because the default response was nil, therefore
unable to detect if the user was seeing the question for the first time
or hadn't previously responded. Now the the default response is an empty
string.