Fix change answer link for query parameter flows
The links to change answers for previous question did not work for query
parameter based flow. This is because previous responses were not saved
in the query string. This mean that changing your answer to a specific
question would take you back to the start of the flow. This updates the
helper function to generate change answer links to include previous
responses in the query string.