Updade custom results tracking action value
We have 3 levels of information on the custom result page: group, topic, and result. We initially setup tracking values as `[topic_index].[result_index]`, but our PA finds it more useful to be `[group_index].[result_index]` so we're updating accordingly. To avoid confusion we explicitly name indexes at the component level and update the result-section component to render a set of sections to maintain component isolation.