Refactor RSpec unit tests to Minitest
This is apart of migration of our existing RSpec testing to Minitest. We
are consollidating our tests to a single testing framework as it was
unclear which framework to write new tests in. These test we're written
in RSpec in attempt to start a migration to RSpec, however now we've
decided that the effort in migrating the existing MiniTest tests is not
worth the benefits we gain from using RSpec. Particular concerned with
the lack of parallel testing which would cause a significant slow down.