Increase group spacing
Increase the group spacing to 80px on desktop (50px on mobile). As the spacing scale currently supports up to 8 units [1] we're going to compose this value. We previously renamed `result-section` component to `result-sections` to accommodate analytics and design needs – so we refactor the SCSS to be consistent with the name