The story so far
It looks like the leak between tests is happening because the tests run
in parallel.
If we set this value: https://github.com/alphagov/smart-answers/blob/main/test/test_helper.rb#L40
to 1, the cache is cleared after every test as expected and all of the
tests pass.

However, CountryAndDateQuestionsTest and ChangingAnswerTest both visit
"/country-and-date-sample/y" and I assume are using the same cache.
When the tests are run in parallel, some of the runs of the
CountryAndDateQuestionsTest are picking up the cache for
ChangingAnswerTest and failing. This is because
CountryAndDateQuestionsTest tests the content of the WorldLocations
drop-down, but ChangingAnswerTest doesn't

I assume the reason that the tests are run in parallel is because the
smart-answers test suite is so slow, and this was an attempt to speed
them up.

One other thing I've tried is using the Rails cache to store the
WorldLocations. The benefit of this is that we can use Rails.cache.fetch
rather than the custom cache_fetch method. We also don't need the
reset_cache method as values in the tests aren't cached
(https://github.com/alphagov/smart-answers/blob/main/config/environments/test.rb#L24-L25).
This also means that we don't have to explicitly tests that the caching
works, Rails should be doing that for us.

If we use the Rails cache, we need to run Rails.cache.clear before each
test, though I'm not sure why if the data isn't cached. Hmmm.

We also have the problem of needing to find a way to use a stale cache
if we can't populate WorldLocations from gds-api-adapters for whatever
reason.