Don't set cache expiry time in tests
There is an integration test that fails intermittently:

```
Failure:
CountryAndDateQuestionsTest#test_: with javascript should handle country and date questions 42.  [/govuk/smart-answers/test/integration/engine/country_and_date_questions_test.rb:111]:
--- expected
+++ actual
@@ -1 +1 @@
-["angola", "aruba", "bangladesh", "belarus", "brazil", "brunei", "cambodia", "chad", "croatia", "denmark", "eritrea", "france", "ghana", "iceland", "japan", "laos", "luxembourg", "malta", "micronesia", "mozambique", "nicaragua", "panama", "portugal", "sao-tome-and-principe", "singapore", "south-korea", "sri-lanka", "venezuela", "vietnam"]
+["argentina", "belarus"]

rails test test/integration/engine/country_and_date_questions_test.rb:43
```

A loop was added [1] to replicate the problem locally.

`CountryAndDateQuestionsTest` and `ChangingAnswerTest` both visit
"/country-and-date-sample/y" but only `CountryAndDateQuestionsTest`
tests the contents of the WorldLocation drop-down, so it's the only test
that fails.

I originally thought this was a caching issue, caused by the tests
running in parallel [2].

After a bit of investigation [3], it turns out it's failing because the
Chrome instance of Capybara is caching a HTTP response for
"/country-and-date-sample/y" (due to expiry headers set).

There were a few ways we could fix this. One of them is was to remove
the cache expiry setting for the Javascript Capybara tests. However,
there are only one set of tests that explicitly test the cache header
[4], so rather turn off the cache expiry for the integration tests, I
have decided to put back the original config that was removed when these
tests were added [4], and stub the cache expiry in the only place that
the cache headers are being tested.

[1]: https://github.com/alphagov/smart-answers/pull/5384/commits/220a6955fef83bfc8b975f6f7933647c0eff682f
[2]: https://github.com/alphagov/smart-answers/blob/main/test/test_helper.rb#L40
[3]: https://github.com/alphagov/smart-answers/commit/83478e0b4f9ec255f753679d7c2a752903fcd542
[4]: https://github.com/alphagov/smart-answers/commit/047df59818126fce8bdb86525a42bf7f4cd1bbe5#diff-3180c288738612297a7a24d405fbade1c6766b27d7050eeee2ecdbef2e14cef0L2