SPIKE - Show proposed test approach
This adds the proposed test layout for two flows. Towing rules is currently a minitest whilst the holiday entitlement was moved over to Rspec. Existing minitest tests are concerned with testing the flow of a question, an approach that was copied when some test suites were moved over to RSpec. Our proposed approach instead focuses on nodal testing. We test each start page, question page and outcome page separately and the assertions focus on which nodes follow which when given a specific history. The major benefit of this is that we greatly reduce the amount of nesting required to test an entire flow. Within the tests we will load up each nodes erb file without rendering the page in a browser. This ensures that syntax errors on the frontend are caught without needing something like capybara and the time that takes to run.

Slight caveat for this particular piece of work is that some of the methods used (such as `see_question`) don't actually exist yet. They will obviously need to be written before these tests could work. This is more about getting feedback on the overall approach of node testing over flow testing.

Many thanks and such!

Co-authored-by: Graham Lewis <graham.lewis@digital.cabinet-office.gov.uk>