Resolve date inconsistencies to fix failing tests
A number of these tests were failing because they used a slash in a
response - which meant each segment of the date was considered a
separate response and caused a failure.

While fixing this I also fixed other oddities around the add_response
arguments in this file so that they're more consistent and easier to
read.