Rename State#current_node to State#current_node_name
This method was quite confusing as you would anticipate a method of
current_node to return a Node object rather than a symbol that
represents the node name. That is the typical  behaviour of other node
methods in this codebase.

I don't think I was quite able to clear up all the confusion surrounding
this method as it still seems surprising to me that this return a symbol
rather than a string (which is the type I'd anticipate a name to occupy),
however this doesn't seem resolvable due to Node#name setting this precedence
and the frequent usage of symbols in node comparisons.

I did first experiment whether it was possible for the #current_node
method to return a node object rather than a symbol as I felt that this
was more idiomatic - and also somewhat sidestepped some of the #name
confusion. I found out, however, that there are various tests that rely on
the ability to return a node which does not exist, so unfortunately that
wasn't a viable path without substantial refactoring.