Update tests for State#current_node_name
This changes all the various references to current_node to the name of
current_node_name.

I did consider whether to rename the frequent assertion
`assert_current_node` to be `assert_current_node_name`, I decided
against this because this assertion tests more than the node name,
though it does seem to be caught a little between two worlds now.