Update hardcoded "London" string in the benefit cap calculator flow
The benefit cap calculator flow currently tests if the input postcode is within
London by testing for a result["type"] of EUR and a result["name"] of London.

MySociety have changed the name of the London area in MapIt from London to
London English Region. At least we thought it had been renamed. Actually it had
been moved to a different GSS code. This exposed the potential for breaking this
code in two ways:

1. Moving a local authority to a different area - changing the GSS code; and
2. Renaming the London area

As a result, the benefit cap calculator's london? method no longer worked
correctly. For example, it returns false for postcodes within Westminister.

This change updates the london? method in the flow's calculator to use a regex
that checks the name for any presence of the word London.

This should buffer us from any change, so long as London is somewhere in the
name of the London area.