Prototype implementation of flow test syntaxes
This serves as an exploratory implementation of the ideas for the node
based flow testing of a Smart Answer Flow.

It has two working flows supported by ExperimentalFlowTestHelper with
various assertions. Don't worry about the state of the code in the
helper, it was put together as a proof of concept rapidly.

Some learnings from putting together this work:

It's quite easy to accidentally get a question name wrong. I've found it
was useful to have something that verifies a question is valid before
setting them.

It's frustrating when you're trying to test something but the set-up
questions fails (for example you've missed a required response). I've
found it beneficial to have error checks prior to assertions to confirm
the expected state - to support this we specify the node under test in
the setup.

It can be quite powerful to have rich assertion functions that check
state while doing the assertions. This seems to lead to relatively easy
to debug code (because of very specific errors) but they could prove a
little magical. I think they should be ok just for the sheer amount of
repetition across Smart Answers.

The shoulda matchers were pretty easy to put together and nice to
write, however they have a pain point that they're hard to run in
isolation when they fail (as they don't have a proper line number) so I
gradually shied away from them and will look at removing them in the
next iteration.

It's confusing having concepts of current and next for a flow state as
they only have one set of responses - what is current in one instance is
next in another. This is an pre-existing challenge and I couldn't think of
how to resolve it without massive duplications so it seems like
something to live with.