Update showing criteria for r28
In the "Next steps for you business" flow, we have a result for "How and
when to charge VAT". This updates the logic so the result is only shown
to business who earn more than £85K, as this result is shown in the
"rules to follow" and isn't a must for business who earn less.