Port basic prefilling of questions for state#current_response
When FlowPresenter was adapted to work directly with a state object the
previous method to access the current response
(response_for_current_question) was removed due to lack of
compatibility. Instead we can access the current_response method from
the flow_presenter which will return the input of the current response.

With the removal of #response_for_current_question there is no longer a
use for the #to_response method so this has been removed.