Port checkbox question for State#current_response
This adapts the code involved in rendering a checkbox question for using
the State#current_response method to establish whether a question is
checked. Unlike the previous iteration this moves all the response
format handling into a single method, where previously it was handled
across the application. The input into the `checked?` method will vary
in type depending on the scenario it is called from (it is most likely
to be a string or an array of strings) and is based on user input, thus
the code is quite defensive in nature.

It also corrects that the #checked? method is defined as public
despite only private usage.