Port radio question for State#current_response
This adapts the code involved in rendering a radio question for using
the State#current_response method to establish whether an option is
selected.

This code is a little bit simpler as we don't need the guard clause
since we don't expect value to be something that returns true to
`blank?`.