Port salary question for State#current_response
This adapts the code involved in rendering a salary question for using
the State#current_response method. Unlike the previous implementation
the code to determine whether a response is checked can cope with a wide
variety of types - this hopefully collates complexity into a single
place.

I decided it was better to have a single method that returned a hash
rather than a collection of methods. This made the code much easier to
test as each permutation of input could be tested once - I don't think
this affects readability negatively.