Consolidate links about Smart Answer development
While the documentation is great, it's also quite niche and is too
much for someone new to the app to take in. This preserves the main
links to develop a new Smart Answer, with a link to the remaining
guidance. Note that the section about publishing Smart Answers is
redundant - it's covered by the publishing.md manual.