Move single doc on testing into further docs
Previously it was unclear what the different between all the other
docs and "app development" was, especially since the guidance on
testing pertained to individual Smart Answers. This preserves the
folder, since we may have more guidance on testing in future, but
there's no value in listing it in its own section.